# docs/swagger.yaml
basePath: /
host: localhost:5000
info:
  contact:
    name: API Support
  description: swagger Projeto Raul
  title: Fiber Example API
  version: "1.0"
paths:
...
  /deleteTerrain/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Terrain by its ID
      operationId: delete-terrain-by-int
      parameters:
      - description: Terrain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Terrain
      tags:
      - Terrain
  /deleteVegetation/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Vegetation by its ID
      operationId: delete-vegetation-by-int
      parameters:
      - description: Vegetation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Vegetation
      tags:
      - Vegetation
  /deleteWaterBody/{id}:
    delete:
      consumes:
      - application/json
      description: delete a WaterBody by its ID
      operationId: delete-waterBody-by-int
      parameters:
      - description: WaterBody ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a WaterBody
      tags:
      - WaterBody
...
  /terrain:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Terrain
  /terrains:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Terrain
...
  /updateTerrain/{id}:
    put:
      consumes:
      - application/json
      description: update a Terrain by its ID
      operationId: get-Terrain-by-int
      parameters:
      - description: Terrain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Terrain
      tags:
      - Terrain
  /updateVegetation/{id}:
    put:
      consumes:
      - application/json
      description: update a Vegetation by its ID
      operationId: get-Vegetation-by-int
      parameters:
      - description: Vegetation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Vegetation
      tags:
      - Vegetation
  /updateWaterBody/{id}:
    put:
      consumes:
      - application/json
      description: update a WaterBody by its ID
      operationId: get-WaterBody-by-int
      parameters:
      - description: WaterBody ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a WaterBody
      tags:
      - WaterBody
  /vegetation:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Vegetation
  /vegetations:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Vegetation
  /waterBody:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new WaterBody
  /waterbodies:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of WaterBody
swagger: "2.0"